refactor: pin 상수 constant 폴더로 옮김
